Changes 1

+++ Auto_Switch/module.php
@@ -773,7 +773,7 @@ private function CreateTimeEvent($ident, $parentID, $Position, $content){
     $idf=IPS_GetEventIDByName('Clear_1', $CatID);
     $ids=IPS_GetEventIDByName('Set_1', $CatID);
     IPS_LogMessage("AutoSwitch_AutoTimeUpdate","EventActive = "
-                .IPS_GetEvent($ids)['EventActive']);
+                .(int)IPS_GetEvent($ids)['EventActive']);
     if(IPS_GetEvent($ids)['EventActive']){
         IPS_LogMessage("AutoSwitch_AutoTimeUpdate","Event = "
                 .$idf." Zeit = ".$Stunde.":".$Minute.":".$Sekunde);
